Add isvhsopen.com Web API support
isvhsopen.com has been rewritten and now has its own API, so we are not
using api.hackspace.ca anymore for the door status. The vhsapi code is
still included because we'll continue to write the RPi's local IP to the
VHS API, to make it easy to find, and the timestamp can serve as a boot
log.